requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
croniter==6.0.0
cssselect==1.4.0
distro==1.9.0
fastapi==0.128.7
fastuuid==0.14.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark-oapi==1.5.3
litellm==1.81.10
loguru==0.7.3
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
# Editable install with no version control (nanobot-ai==0.1.3.post5)
-e Agent/nanobot
openai==2.20.0
packaging==26.0
propcache==0.4.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
pytz==2025.2
PyYAML==6.0.3
readability-lxml==0.8.4.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
starlette==0.52.1
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.22.0
typer-slim==0.22.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websocket-client==1.9.0
websockets==16.0
win32_setctime==1.2.0
yarl==1.22.0
zipp==3.23.0
nanobot-ai